fix for context derivation of par/gtx flags
As reported in ticket #291, when sig_flag is inferred, the variables
m_tmplCpDiag and m_tmplCpSum1 are not set correctly, but used for
deriving the contexts for par_glag, gt1_flag, and gt3_flag.

The provided modification fixes this issue by calling cctx.sigCtxIdAbs()
also when sig_flag is inferred (except for the last scanning position,
where a different mechanism is used).

There is no impact on coding efficiency.