Update chroma QP tables for use in CTC
This should implement the following from the meeting notes:

For CTC, it was initially agreed to use the new syntax to achieve
the chroma QP settings proposed in O0186 for SDR, N0221 for PQ,
O0433 for HLG. For the joint table, use the average of the Cb and Cr
tables. Below QP = 16, chroma and luma QP should be equal for CEs
using low QP.

Decision (CTC): This was further discussed in the closing plenary
on Friday. It was suggested to keep the anchor QP settings for PQ
and HLG as in CTC of prior meeting cycle. For SDR, see above.
Below QP = 16, chroma and luma QP should be equal for CEs using low
QP.